Create ProfileWorld component using World3D helpers
