+++ vite-react-main/src/components/Shell.tsx
@@ -1,4 +1,4 @@
-import { useMemo } from "react";
+import { useNavigate } from "react-router-dom";
 import "./Shell.css";
 import type { Post, User } from "../types";
 import BrandBadge from "./BrandBadge";
@@ -6,29 +6,13 @@ import PostCard from "./PostCard";
 import AssistantOrb from "./AssistantOrb";
 import World3D from "./World3D";
 
-const IMG = (id: number) => `https://picsum.photos/id/${id}/1080/1350`;
-const SEED = [
-  1015,1069,1025,1027,1043,1050,106,237,1005,1003,1002,1001,
-  1021,1024,1032,1035,1037,1040,1041,1042,1044,1045,1049,1051,
-  1055,1056,1059,1063,1067,1070,
-];
+type Props = {
+  posts: Post[];
+  me: User;
+};
 
-export default function Shell() {
-  const avatar = (i: number) => `https://i.pravatar.cc/100?img=${(i % 70) + 1}`;
-  const me: User = { id: "me", name: "You", avatar: avatar(99) };
-
-  const posts: Post[] = useMemo(
-    () =>
-      SEED.map((id, i) => ({
-        id: `p${i}`,
-        author: i % 3 === 0 ? "Elena R." : i % 3 === 1 ? "Expanso" : "Lola",
-        authorAvatar: avatar(i),
-        title: i % 5 === 0 ? "Travel" : undefined,
-        time: `${(i % 8) + 1}h`,
-        images: [{ id: `i${i}`, url: IMG(id) }],
-      })),
-    []
-  );
+export default function Shell({ posts, me }: Props) {
+  const navigate = useNavigate();
 
   const onEnterWorld = () => {
     console.log("Enter Universe");
@@ -52,7 +36,7 @@ export default function Shell() {
               key={p.id}
               post={p}
               me={me}
-              onOpenProfile={(id) => console.log("profile:", id)}
+              onOpenProfile={(id) => navigate(`/profile/${id}`)}
               onEnterWorld={onEnterWorld}
             />
           ))}
