+++ vite-react-main/src/components/index.ts
@@ -1,5 +1,6 @@
 export { default as Sidebar } from "./Sidebar";
 export { default as Feed2D } from "./Feed2D";
 export { default as World3D } from "./World3D";
+export { default as ProfileWorld } from "./ProfileWorld";
 export { default as AssistantOrb } from "./AssistantOrb";
 export type { Post } from "../types";
